requirements.txt
google-genai
gradio
pandas
gtts
flask
python-dotenv
psycopg2-binary

